+++ src/custom/pages/Swap/components/ConfirmSwapModalSetup.tsx
@@ -5,7 +5,7 @@ import { Percent } from '@uniswap/sdk-core'
 import { Field } from 'state/swap/actions'
 import { useAtomValue, useUpdateAtom } from 'jotai/utils'
 import { swapConfirmAtom } from 'pages/Swap/state/swapConfirmAtom'
-import { HandleSwapCallback } from 'pages/Swap/helpers/useHandleSwap'
+import { HandleSwapCallback } from 'pages/Swap/hooks/useHandleSwap'
 
 export interface ConfirmSwapModalSetupProps {
   trade: TradeGp | undefined
